added creation of pt directory

+++ install.bash
@@ -23,6 +23,9 @@ then
   ptdir=win	
 fi
 
+test -d src/main/pt || mkdir src/main/pt || die "Could not create pt directory?"
+
+echo "Copying from install/$ptdir/pt to src/main/pt/"
 cp -R install/$ptdir/pt src/main/pt/ || die "Could not copy pluggable transports?"
 
 rm -f target/lantern*-small.jar || die "Could not remove old jar?"
